+++ netlify.toml
@@ -1,0 +1,4 @@
+[build]
+command = "yarn build"
+functions = "netlify_functions"
+publish = "dist"
