--- netlify_functions/submission-created/.env
@@ -1,4 +1,0 @@
-SMTP_HOST=smtp.mail.us-east-1.awsapps.com
-SMTP_PORT=465
-SMTP_USER=fumito.suzuki@transpigeon.com
-SMTP_PASS=rebHo2-datqoh-neztot
