+++ package.json
@@ -21,6 +21,7 @@
     "contentful": "^7.14.8",
     "core-js": "^3.6.5",
     "markdown-it-highlightjs": "3.3.1",
+    "nodemailer": "6.4.16",
     "nuxt": "^2.14.6",
     "nuxt-basic-auth-module": "^1.3.3",
     "prism-themes": "^1.5.0",
